Component should start even without attachment override or active attach.
This is a chicken and the egg problem, prior to this we said we could
not start without an attachment override or attach but an attach
can only happen after a stream has started flowing which generally
happens after a waveform is started. So now we can start, but are
in a waiting state. Needs unit tests!